the sleeptimer dialog was transfered to dialog helper.
